+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   cause undefined behavior when called with a `Layout` with an alignment
   other than 1. A safe alternative is to use
   [`Vec::into_boxed_slice`](https://doc.rust-lang.org/std/vec/struct.Vec.html#method.into_boxed_slice).
+- Removed `From` conversions from `ucs2::Error` to `Status` and `Error`.
 
 ## uefi-macros - [Unreleased]